Fix custom rule when the task does not exist
When the task does not exist, the rule currently ends up
with a NullPointerException. Not only is it misleading but it
also is a problem when you want to run a custom task on all
subprojects of a project: some projects may not have the task
to customize. In that case we want the rule to do nothing

Also upgraded version as 1.0 was already released